regroup listener functions
