--- kuzzle/remove-all-listeners.go
@@ -1,6 +1,0 @@
-package kuzzle
-
-// Remove all listener by event type or all listener if event == -1
-func (k *Kuzzle) RemoveAllListeners(event int) {
-	k.socket.RemoveAllListeners(event)
-}
