--- kuzzle/remove_listener.go
@@ -1,6 +1,0 @@
-package kuzzle
-
-// RemoveListener removes a listener
-func (k *Kuzzle) RemoveListener(event int) {
-	k.socket.RemoveListener(event)
-}
